--- dropEmployeeTable.py
@@ -1,12 +1,0 @@
-import mysql.connector
-
-mydb = mysql.connector.connect(
-    host="localhost",
-    user="tester",
-    password="tester",
-    database="LegoStore"
-)
-
-mycursor = mydb.cursor()
-
-mycursor.execute("DROP TABLE employees")
