+++ employeeTable/createEmployeesTable.py
@@ -1,0 +1,12 @@
+import mysql.connector
+
+mydb = mysql.connector.connect(
+    host="localhost",
+    user="tester",
+    password="tester",
+    database="LegoStore"
+)
+
+mycursor = mydb.cursor()
+
+mycursor.execute("CREATE TABLE employees (id INT AUTO_INCREMENT PRIMARY KEY, name VARCHAR(255), store_preference VARCHAR(255))")
