+++ employeeTable/dropEmployeeTable.py
@@ -1,0 +1,12 @@
+import mysql.connector
+
+mydb = mysql.connector.connect(
+    host="localhost",
+    user="tester",
+    password="tester",
+    database="LegoStore"
+)
+
+mycursor = mydb.cursor()
+
+mycursor.execute("DROP TABLE employees")
